// generated/src/app/api-models/claim-api-othr-carr-w.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ClaimApiOthrCarrW {
  claimNumber: string;
  sequenceNo: string;
  payorId: string;
  providerNumber: string;
  patReleaseInd: string;
  assignBenefitsInd: string;
  priorPayment: string;
  estAmtDue: string;
  insuredLastName: string;
  insuredFirstName: string;
  insuredMi: string;
  patRelToInsured: string;
  idNumber: string;
  groupName: string;
  groupNumber: string;
  treatmentAuthCode: string;
  employmentStatusCode: string;
  insuredEmpName: string;
  insuredEmpLocation: string;
  internalControlNo: string;
}